Infomag-v2 / 27.June.2019 - image function working well

+++ frontend/src/components/landing_create/section_4_term.vue
@@ -7,20 +7,20 @@
 
       <div class="col-sm-9 mt-sm-3">
         <label class="switch" for="term_status">
-          <input type="checkbox" id="term_status" :value="term_switch"
+          <input type="checkbox" id="term_status" v-model="term_state"
                  @change="change_switch('term')">
           <span class="slider round"></span>
         </label>
       </div>
 
-      <label v-if="term_switch" class="col-sm-3 col-form-label-sm mt-3"
+      <label v-if="term_state" class="col-sm-3 col-form-label-sm mt-3"
              for="term_switch">
         약관 이미지
       </label>
 
-      <div v-if="term_switch" class="col-sm-9 mt-sm-3">
+      <div v-if="term_state" class="col-sm-9 mt-sm-3">
         <label class="switch" for="term_switch">
-          <input type="checkbox" id="term_switch" :value="image_switch"
+          <input type="checkbox" id="term_switch" v-model="image_state"
                  @change="change_switch('image')">
           <span class="slider round"></span>
         </label>
@@ -29,17 +29,26 @@
     </div>
 
     <div class="form-group row"
-         v-if="term_switch && image_switch">
+         v-if="term_state && image_state">
       <label class="col-sm-3 col-form-label-sm mt-3" for="term_img">약관 이미지 파일</label>
       <div class="col-sm-9 mt-sm-3 row ml-0">
         <input type="file" class="input_one_btn form-control col-md-11 pt-1" id="term_img" placeholder="이미지"
                ref="term_file_input" @change="term_file_add($event.target.files[0])" accept="image/*">
         <button type="button" class="btn btn-danger col-md-1 p-0" @click.prevent="term_file_delete()">삭제</button>
       </div>
+      <label class="col-sm-3 col-form-label-sm mt-3" for="term_img_preview">약관 이미지 미리보기</label>
+      <div class="col-sm-9 mt-sm-3 row ml-0" id="term_img_preview">
+        <div v-if="term.image_data" class="term_preview_wrap">
+          <img class="term_preview" :src="key_to_url(term.image_data)" alt="약관 이미지 미리보기">
+        </div>
+        <div v-else class="form-control">
+          <div>등록된 파일이 없습니다</div>
+        </div>
+      </div>
     </div>
 
     <div class="form-group row"
-         v-if="term_switch && !image_switch">
+         v-if="term_state && !image_state">
       <label class="col-sm-3 col-form-label-sm mt-3" for="term_title">약관 제목</label>
       <div class="col-sm-9 mt-sm-3">
         <input type="text" class="form-control" id="term_title" placeholder="title"
@@ -68,25 +77,30 @@
       'term',
       'push_landing'
     ],
+    data: () => ({
+      term_state: false,
+      image_state: false
+    }),
     mounted() {
-
+      this.init_switch()
     },
     methods: {
+      init_switch() {
+        this.term_state = this.term_switch
+        this.image_state = this.image_switch
+      },
       change_switch(option) {
-        if (option === 'term') {
-          this.$emit('update:term_switch', !this.term_switch)
+        if (option == 'term') {
+          this.$emit('update:term_switch', this.term_state)
           this.push_landing()
-        } else if (option === 'image') {
-          this.$emit('update:image_switch', !this.image_switch)
+        } else if (option == 'image') {
+          this.$emit('update:image_switch', this.image_state)
           this.push_landing()
         }
 
       },
       term_file_add(file) {
         /* When file data changed */
-        // let file_data = $event.target.files[0]
-        // this.dynamo_obj.landing_info.term.image = file_data
-
         let key = require('../../../vue_env')
 
         AWS.config.update({
@@ -117,7 +131,7 @@
             console.log('S3 method error occurred', error)
           } else {
             // console.log('S3 method success', data)
-            this.term.image = 'assets.infomagazine.xyz/images/landing/preview/' + this.epoch_time + '/term/' + file.lastModified + '_' + file.name
+            this.term.image_data = params.Key
             this.push_landing()
           }
         })
@@ -142,25 +156,63 @@
           }
         )
 
-        let photoKey = 'assets/images/landing/preview/' + this.epoch_time + '/term/' + this.term.image
+        let photoKey = this.term.image_data
 
-        s3.deleteObject({Key: photoKey}, (err, data) => {
-          if (err) {
-            alert('There was an error deleting your photo: ', err.message)
-          } else {
-            // alert('Successfully deleted photo.', data)
-            document.getElementById('term_img').value = ''
-            this.term.image = null
+        if (photoKey) {
+          s3.deleteObject({Key: photoKey}, (err, data) => {
+            if (err) {
+              alert('이미지 삭제 중 오류가 발생하였습니다: ', err.message)
+            } else {
+              // alert('Successfully deleted photo.', data)
+              document.getElementById('term_img').value = ''
+              this.term.image_data = null
 
-            this.push_landing()
-          }
-        })
+              this.push_landing()
+            }
+          })
+        }
 
       },
+      key_to_url(key) {
+        let divided = key.split('/')
+        let url = 'https://'
+
+        if (this.updated_date == '') {
+          url += 'infomagazine.s3.ap-northeast-2.amazonaws.com/' + key
+        } else {
+          for (let i = 0; i < divided.length; i++) {
+            if (i == 0) {
+              url += (divided[i] + '.infomagazine.xyz')
+            } else {
+              url += ('/' + divided[i])
+            }
+          }
+        }
+
+        return url
+      }
     }
   }
 </script>
 
-<style scoped>
+<style lang="scss" scoped>
+  .term_preview_wrap {
+    display: block;
+    width: 100%;
+    padding: .375rem .75rem;
+    font-size: 1rem;
+    line-height: 1.5;
+    color: #495057;
+    background-color: #fff;
+    background-clip: padding-box;
+    border: 1px solid #ced4da;
+    border-radius: .25rem;
+    transition: border-color .15s ease-in-out, box-shadow .15s ease-in-out;
+    text-align: center;
+  }
 
+  .term_preview {
+    position: relative;
+    max-width: 100%;
+  }
 </style>
